+++ changelog.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Fix: Home 2 Section 1 removed space-between and replaced it by a responsive Stack
 - Fix: Navigation menu with light bg when in respective route
 - Fix: Removed mock data for testimonials
+- Fix: Fixed slidy testimonials user picture into home 2
 
 ### [1.30.2] - 2022-11-14